requirements.txt
mysql-connector-python==8.1.0
pika==1.3.2
prometheus-client==0.17.1
protobuf==4.21.12
psycopg2-binary==2.9.9
PyYAML==6.0.1
redis==5.0.1
